Quick change before work
I got back late so i have tto dip out for work rn, but i only made it so the ninja ran a little smoother on the platform, and made him turn gold for immunity, Just need to make it so when he is gold he doesn't just contiously do the same animation over and stop doing it if you try to run with it on. I should be able to figure it out by tonihgt.